Move compiler warnings settings to target level for modules (#479)
* Apply the warnings to modules instead of globally

* Consolidates warnings into one list so we can do one CI_BUILD to check our builds locally

* Moves the compiler warnings function into nuclear/cmake

* Change camera config field names back to PascalCase (#480)

* Code review - Alex's suggestions

* Changes default WARNINGS_AS_ERRORS setting

* Added compiler warnings for messages, extensions, and utilities

* Fixes formatting on d.ts files (#483)

* Moves extension and utility warning application to their libraries.cmake files

* Removes $NUBOTS_ROOT/cmake from the module path, because it wasn't needed any more

* formatting lel

* Revert "Merge remote-tracking branch 'origin/master' into hamiltons/compiler_warnings_changes"

This reverts commit e5bff3b8170aa75f065977815adb41f6029dd450.

* Revert "Merge remote-tracking branch 'origin/master' into biddulph/compiler-warnings"

This reverts commit e09d638ec2d2dfaab0131215220b9725bfea4fab.

* Adds warnings for module and utility tests

* Removes redundant module path additions

* Formatting

* Removes warnings from conversiontest

* Removes shadowed variable in FileWatcherTest

* Fixes test warnings

* Removes gcc-only warnings clang-tidy gets

* Deletes function file for warnings

* Moves warnings to main cmake file with configurable variable

* Moves warning setting into nuclear directory, with separate warnings for each component

* Formatting

* Suggestion from alex I forgot to push

* Refines the setting of the warnings variables

* Introduces many casts so we do signed arithmetic without warnings

* More signedness changes on TestSplines

* Removes bad warning which warns on int indices

* Responds to trent's review

* Formatting

* Screaming snakes for shadowed var

* Makes constants screaming snake case

* Makes servo indices ints in `Convert`

+++ module/platform/darwin/HardwareIO/src/Convert.cpp
@@ -80,14 +80,14 @@ namespace module::platform::darwin {
                                : uint8_t(std::round(value / GAIN_CONVERSION_FACTOR));  // Otherwise do our conversion
     }
 
-    float Convert::servoPosition(const uint8_t& id, const uint16_t& value) {
+    float Convert::servoPosition(const int& id, const uint16_t& value) {
         // offset and normalize the angle
         return utility::math::angle::normalizeAngle(float(value - 2048) * POSITION_CONVERSION_FACTOR)
                    * float(SERVO_DIRECTION[id])
                + SERVO_OFFSET[id];
     }
 
-    uint16_t Convert::servoPositionInverse(const uint8_t& id, const float& value) {
+    uint16_t Convert::servoPositionInverse(const int& id, const float& value) {
         float angle = value;
 
         // Undo our conversion operations
@@ -101,7 +101,7 @@ namespace module::platform::darwin {
         return uint16_t(angle / POSITION_CONVERSION_FACTOR) + 2048;
     }
 
-    float Convert::servoSpeed(const uint8_t& id, const uint16_t& value) {
+    float Convert::servoSpeed(const int& id, const uint16_t& value) {
 
         // We only care about the lower bits
         float raw = float(value & 0x3FF) * SPEED_CONVERSION_FACTOR;
@@ -127,7 +127,7 @@ namespace module::platform::darwin {
         return value >= 100.0f ? 1023 : uint16_t(std::round(value / TORQUE_LIMIT_CONVERSION_FACTOR));
     }
 
-    float Convert::servoLoad(const uint8_t& id, const uint16_t& value) {
+    float Convert::servoLoad(const int& id, const uint16_t& value) {
         // We only care about the lower bits if bit 10 is set then we are moving clockwise
         float raw = float(value & 0x3FF) * (value & 0x400 ? -1.0f : 1.0f);
         raw *= LOAD_CONVERSION_FACTOR;
